pyproject.toml
[project]
name = "coinpursefour"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "sqlalchemy>=2.0.36",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.8.3",
]
